Implement custom roles for scheduled events (#23)
* setup drizzle orm

* implement mysql manager

The mysql connection will initialize and end itself with the client.

* create custom role on new schedule event creation

* write queries for event-role feature

* delete custom roles on scheduled event completion or deletion

* enable partials for scheduled events and guild management

* changed how logs are formatted

* implement custom role assignment on enrolling or dropping from a scheduled event

* implement a scheduled event enrollment queue

On occasion the call to assign the custom role to the scheduled event
author happens before the database entry is created, causing a race
condition. The enrollment queue prevents that.

* write documention

* create a function that truncates strings

It truncates strings to a specified number of words (delimited by
spaces) or a character count limit, which ever hits first.